+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/models/StorageAccountInfo.java
@@ -11,6 +11,7 @@ import com.azure.core.annotation.Immutable;
 public class StorageAccountInfo {
     private final SkuName skuName;
     private final AccountKind accountKind;
+    private final Boolean isHnsEnabled;
 
     /**
      * Constructs a {@link StorageAccountInfo}.
@@ -21,6 +22,20 @@ public class StorageAccountInfo {
     public StorageAccountInfo(final SkuName skuName, final AccountKind accountKind) {
         this.skuName = skuName;
         this.accountKind = accountKind;
+        this.isHnsEnabled = null;
+    }
+
+    /**
+     * Constructs a {@link StorageAccountInfo}.
+     *
+     * @param skuName SKU of the account.
+     * @param accountKind Type of the account.
+     * @param isHnsEnabled whether hierarchical namespace is enabled on the account.
+     */
+    public StorageAccountInfo(final SkuName skuName, final AccountKind accountKind, Boolean isHnsEnabled) {
+        this.skuName = skuName;
+        this.accountKind = accountKind;
+        this.isHnsEnabled = isHnsEnabled;
     }
 
     /**
@@ -35,5 +50,12 @@ public class StorageAccountInfo {
      */
     public AccountKind getAccountKind() {
         return accountKind;
+    }
+
+    /**
+     * @return whether hierarchical namespace is enabled on the account.
+     */
+    public Boolean isHnsEnabled() {
+        return isHnsEnabled;
     }
 }